Add dialog when migration error occurs
